requirements.txt
ballpark==1.4.0
einops==0.3.0
h5py==3.7.0
hydra-core==1.3.1
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgaug==0.2.6
imgviz==1.6.2
lightning-utilities==0.5.0
matplotlib==3.5.3
matplotlib-inline==0.1.6
numpy==1.23.3
omegaconf
opencv-python==4.6.0.66
Pillow==9.2.0
plotly==5.7.0
pypng==0.20220715.0
pytorch-lightning==1.8.6
rich==12.5.1
scikit-image==0.19.3
scipy==1.9.1
tabulate==0.9.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
termcolor==2.2.0
torch==2.2.2
torch-efficient-distloss==0.1.3
torch-fidelity==0.3.0
#torch-scatter -f https://data.pyg.org/whl/torch-2.2.2+cu121.html
torchaudio
torchmetrics==0.11.0
torchtyping==0.1.4
torchvision
tornado==6.2
tqdm==4.64.1
transforms3d==0.4.1
trimesh==3.18.0
wandb==0.13.4
